@@ -7,8 +7,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.1
 
 environment:
-  sdk: ">=2.14.0-0 <3.0.0"
-  flutter: ">=2.8.1-0 <3.0.0"
+  sdk: ">=3.0.0 < 4.0.0"
 
 dependencies:
   flutter:
@@ -16,44 +15,43 @@ dependencies:
   dartz: ^0.10.1
   enum_to_string: ^2.0.1
   file: ^7.0.0
-  cross_file: ^0.3.2
-  flutter_archive: ^6.0.0
+  cross_file: ^0.3.4+1
+  flutter_archive: ^6.0.3
   freezed_annotation: ^2.4.1
   internet_connection_checker: ^1.0.0+1
-  json_annotation: ^4.8.1
-  drift: ^2.16.0
-  sqlite3_flutter_libs: ^0.5.2
-  path: ^1.8.1
-  path_provider: ^2.0.7
-  drift_db_viewer: ^2.0.0
+  json_annotation: ^4.9.0
+  drift: ^2.18.0
+  sqlite3_flutter_libs: ^0.5.21
+  path: ^1.9.0
+  path_provider: ^2.1.3
+  drift_db_viewer: ^2.1.0
   android_path_provider: ^0.3.0
-  reactive_forms: ^16.1.1
-  hooks_riverpod: ^ 2.5.1
+  reactive_forms: ^17.0.1
+  hooks_riverpod: ^2.5.1
   #reactive_multi_image_picker:
     #git:
       #url: https://github.com/victorcatn/reactive_multi_image_picker.git
       #ref: multi_platform_support
-  rflutter_alert: ^2.0.4
+  rflutter_alert: ^2.0.7
   sentry_flutter: ^8.1.0
-  shared_preferences: ^2.0.11
-  simple_tooltip: ^1.0.0
-  url_launcher: ^6.0.17
+  shared_preferences: ^2.2.3
+  simple_tooltip: ^1.2.0
+  url_launcher: ^6.2.6
   go_router: ^14.1.0
-  connectivity_plus: 5.0.2
-  uuid: ^4.3.3
-  rxdart: ^0.27.3
-  diacritic: ^0.1.3
+  connectivity_plus: ^6.0.3
+  uuid: ^4.4.0
+  rxdart: ^0.27.7
+  diacritic: ^0.1.5
 
 dev_dependencies:
-  build_runner: ^2.4.8
-  drift_dev: ^2.16.0
+  build_runner: ^2.4.9
+  drift_dev: ^2.18.0
   flutter_launcher_icons: ^0.13.1
-  freezed: ^2.4.7
-  json_serializable: ^6.7.1
-  flutter_lints: ^3.0.1
-  test: ^1.25.2
+  freezed: ^2.5.2
+  json_serializable: ^6.8.0
+  flutter_lints: ^4.0.0
+  test: ^1.25.5
   mockito: ^5.4.4
-  file_testing: ^3.0.0
   
 
 flutter_icons:
